# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SerialMemory API",
    "description": "Temporal knowledge graph memory system with semantic search, entity extraction, and multi-hop reasoning.",
    "contact": {
      "name": "SerialMemory",
      "url": "https://serialmemory.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/admin/control-room/actions": {
      "get": {
        "tags": [
          "Control Room"
        ],
        "description": "Gets recent kill switch actions for audit trail",
        "operationId": "GetControlRoomActions",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KillSwitchAction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/control-room/global/kill": {
      "post": {
        "tags": [
          "Control Room"
        ],
        "description": "Enables the global kill switch - blocks ALL requests (503)",
        "operationId": "EnableGlobalKillSwitch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Control Room"
        ],
        "description": "Disables the global kill switch",
        "operationId": "DisableGlobalKillSwitch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiftKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/control-room/global/readonly": {
      "post": {
        "tags": [
          "Control Room"
        ],
        "description": "Enables global read-only mode - allows GET, blocks write verbs",
        "operationId": "EnableGlobalReadOnly",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Control Room"
        ],
        "description": "Disables global read-only mode",
        "operationId": "DisableGlobalReadOnly",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiftKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/control-room/tenant/{tenantId}/kill": {
      "post": {
        "tags": [
          "Control Room"
        ],
        "description": "Disables all operations for a specific tenant",
        "operationId": "EnableTenantKillSwitch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Control Room"
        ],
        "description": "Re-enables a disabled tenant",
        "operationId": "DisableTenantKillSwitch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiftKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/control-room/tenant/{tenantId}/readonly": {
      "post": {
        "tags": [
          "Control Room"
        ],
        "description": "Sets a tenant to read-only mode",
        "operationId": "EnableTenantReadOnly",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Control Room"
        ],
        "description": "Disables read-only mode for a tenant",
        "operationId": "DisableTenantReadOnly",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiftKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/control-room/tenant/{tenantId}/no-ingest": {
      "post": {
        "tags": [
          "Control Room"
        ],
        "description": "Disables memory ingestion for a tenant (searches still work)",
        "operationId": "EnableTenantNoIngest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Control Room"
        ],
        "description": "Re-enables memory ingestion for a tenant",
        "operationId": "DisableTenantNoIngest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiftKillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/control-room/api-key/{keyId}/disable": {
      "post": {
        "tags": [
          "Control Room"
        ],
        "description": "Disables a specific API key",
        "operationId": "DisableApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/control-room/api-key/{keyId}/enable": {
      "post": {
        "tags": [
          "Control Room"
        ],
        "description": "Re-enables a disabled API key",
        "operationId": "EnableApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/control-room/search/tenant": {
      "get": {
        "tags": [
          "Control Room"
        ],
        "description": "Search for tenants by name, slug, or email",
        "operationId": "SearchTenants",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantSearchResult"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyActionRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GlobalKillSwitchRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          },
          "expiresInMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "KillSwitchAction": {
        "required": [
          "actionType",
          "id",
          "performedAt",
          "performedBy",
          "reason",
          "targetType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "actionType": {
            "type": "string",
            "nullable": true
          },
          "targetType": {
            "type": "string",
            "nullable": true
          },
          "targetId": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "performedBy": {
            "type": "string",
            "nullable": true
          },
          "performedAt": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LiftKillSwitchRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TenantKillSwitchRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          },
          "expiresInMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TenantSearchResult": {
        "required": [
          "createdAt",
          "name",
          "slug",
          "status",
          "tenantId"
        ],
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API key for authentication",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}